Refactored global repositories into their own proper namespaces
