+++ module/Core/src/Crawling/CrawlingHelper.php
@@ -6,7 +6,7 @@ namespace Shlinkio\Shlink\Core\Crawling;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
-use Shlinkio\Shlink\Core\Repository\ShortUrlRepositoryInterface;
+use Shlinkio\Shlink\Core\ShortUrl\Repository\ShortUrlRepositoryInterface;
 
 class CrawlingHelper implements CrawlingHelperInterface
 {
