+++ module/Core/test/Tag/Paginator/Adapter/TagsInfoPaginatorAdapterTest.php
@@ -8,9 +8,9 @@ use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
-use Shlinkio\Shlink\Core\Repository\TagRepositoryInterface;
 use Shlinkio\Shlink\Core\Tag\Model\TagsParams;
 use Shlinkio\Shlink\Core\Tag\Paginator\Adapter\TagsInfoPaginatorAdapter;
+use Shlinkio\Shlink\Core\Tag\Repository\TagRepositoryInterface;
 
 class TagsInfoPaginatorAdapterTest extends TestCase
 {
